+++ Multisig/Logic/Models/HashString.swift
@@ -7,7 +7,6 @@
 //
 
 import Foundation
-import UIKit
 
 struct HashString: Hashable, Codable {
     let hash: Data
@@ -42,24 +41,5 @@ struct HashString: Hashable, Codable {
 extension HashString: CustomStringConvertible {
     var description: String {
         hash.toHexStringWithPrefix()
-    }
-}
-
-extension HashString {
-    var highlighted: NSAttributedString {
-        let style = GNOTextStyle.tertiary
-        let attributedString = NSMutableAttributedString(
-            string: description,
-            attributes: [
-                .font: UIFont.gnoFont(forTextStyle: style),
-                .foregroundColor: style.color!,
-                .kern: -0.41
-            ]
-        )
-        attributedString.addAttribute(
-            .foregroundColor, value: UIColor.primaryLabel, range: NSRange(location: 0, length: 6))
-        attributedString.addAttribute(
-            .foregroundColor, value: UIColor.primaryLabel, range: NSRange(location: description.count - 4, length: 4))
-        return attributedString
     }
 }
